Homework for Lesson 2.4 (correction version of dependencies)
