+++ src/main/java/ru/geekbrains/spring/market/endpoints/ProductEndpoint.java
@@ -6,9 +6,9 @@ import org.springframework.ws.server.endpoint.annotation.PayloadRoot;
 import org.springframework.ws.server.endpoint.annotation.RequestPayload;
 import org.springframework.ws.server.endpoint.annotation.ResponsePayload;
 import ru.geekbrains.spring.market.services.ProductService;
-import ru.geekbrains.spring.market.soap.GetAllProductsResponse;
-import ru.geekbrains.spring.market.soap.GetProductByIdRequest;
-import ru.geekbrains.spring.market.soap.GetProductByIdResponse;
+import ru.geekbrains.spring.market.soap.products.GetAllProductsResponse;
+import ru.geekbrains.spring.market.soap.products.GetProductByIdRequest;
+import ru.geekbrains.spring.market.soap.products.GetProductByIdResponse;
 
 @Endpoint
 @RequiredArgsConstructor
@@ -17,6 +17,18 @@ public class ProductEndpoint {
     private static final String NAMESPACE_URI = "http://www.market.ru/spring/ws/products";
     private final ProductService productService;
 
+    /*
+        Пример запроса: POST http://localhost:8189/market/ws
+
+        <soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:ps="http://www.market.ru/spring/ws/products">
+            <soapenv:Header/>
+            <soapenv:Body>
+                <ps:getProductByIdRequest>
+                    <ps:id>17</ps:id>
+                </ps:getProductByIdRequest>
+            </soapenv:Body>
+        </soapenv:Envelope>
+     */
     @PayloadRoot(namespace = NAMESPACE_URI, localPart = "getProductByIdRequest")
     @ResponsePayload
     public GetProductByIdResponse getProductById(@RequestPayload GetProductByIdRequest request) {
@@ -26,7 +38,7 @@ public class ProductEndpoint {
     }
 
     /*
-        Пример запроса: POST http://localhost:8189/ws
+        Пример запроса: POST http://localhost:8189/market/ws
 
         <soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:ps="http://www.market.ru/spring/ws/products">
             <soapenv:Header/>
@@ -35,7 +47,6 @@ public class ProductEndpoint {
             </soapenv:Body>
         </soapenv:Envelope>
      */
-
     @PayloadRoot(namespace = NAMESPACE_URI, localPart = "getAllProductsRequest")
     @ResponsePayload
     public GetAllProductsResponse getAllProducts() {
