+++ src/main/java/ru/geekbrains/spring/market/soap/products/GetAllProductsRequest.java
@@ -6,13 +6,12 @@
 //
 
 
-package ru.geekbrains.spring.market.soap;
+package ru.geekbrains.spring.market.soap.products;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
-
 
 /**
  * <p>Java class for anonymous complex type.
