+++ src/main/java/ru/geekbrains/spring/market/soap/products/GetAllProductsResponse.java
@@ -6,7 +6,7 @@
 //
 
 
-package ru.geekbrains.spring.market.soap;
+package ru.geekbrains.spring.market.soap.products;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,7 +15,6 @@ import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
-
 
 /**
  * <p>Java class for anonymous complex type.
