+++ src/main/java/ru/geekbrains/spring/market/soap/products/GetProductByIdResponse.java
@@ -6,14 +6,13 @@
 //
 
 
-package ru.geekbrains.spring.market.soap;
+package ru.geekbrains.spring.market.soap.products;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
-
 
 /**
  * <p>Java class for anonymous complex type.
